+++ common/src/main/java/me/ultrusmods/altorigingui/Constants.java
@@ -1,0 +1,11 @@
+package me.ultrusmods.altorigingui;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class Constants {
+
+	public static final String MOD_ID = "altorigingui";
+	public static final String MOD_NAME = "AltOriginGui";
+	public static final Logger LOG = LoggerFactory.getLogger(MOD_NAME);
+}