+++ fabric/src/main/java/me/ultrusmods/altorigingui/AltOriginGuiFabric.java
@@ -1,0 +1,10 @@
+package me.ultrusmods.altorigingui;
+
+import net.fabricmc.api.ModInitializer;
+
+public class AltOriginGuiFabric implements ModInitializer {
+    
+    @Override
+    public void onInitialize() {
+    }
+}