+++ forge/src/main/java/me/ultrusmods/altorigingui/AltOriginGuiForge.java
@@ -1,0 +1,11 @@
+package me.ultrusmods.altorigingui;
+
+import net.minecraftforge.fml.common.Mod;
+
+@Mod(Constants.MOD_ID)
+public class AltOriginGuiForge {
+    
+    public AltOriginGuiForge() {
+
+    }
+}