+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-- Fix extra blank tabs showing up in some situations.
-- This is a small update to fix a bug, the next update will larger, and add a search bar, and sorting to the GUI.
+- Rewrite mod to multiloader, supporting Origins for fabric & forge.